+++ src/StreamCraft/src/StreamCraft_mzxml.hpp
@@ -39,17 +39,17 @@ namespace sc {
         int extract_spec_level() const;
         int extract_spec_mode() const;
         int extract_spec_polarity() const;
-        double extract_spec_lowmz() const;
-        double extract_spec_highmz() const;
-        double extract_spec_bpmz() const;
-        double extract_spec_bpint() const;
-        double extract_spec_tic() const;
-        double extract_scan_rt() const;
-        double extract_ion_mz() const;
-        double extract_activation_ce() const;
+        float extract_spec_lowmz() const;
+        float extract_spec_highmz() const;
+        float extract_spec_bpmz() const;
+        float extract_spec_bpint() const;
+        float extract_spec_tic() const;
+        float extract_scan_rt() const;
+        float extract_ion_mz() const;
+        float extract_activation_ce() const;
         bool has_precursor() const { return spec.child("precursorMz"); }
         MZXML_BINARY_METADATA extract_binary_metadata() const;
-        std::vector<std::vector<double>> extract_binary_data(const MZXML_BINARY_METADATA&  mtd) const;
+        std::vector<std::vector<float>> extract_binary_data(const MZXML_BINARY_METADATA&  mtd) const;
       
       private:
         const pugi::xml_node& spec;
@@ -92,34 +92,34 @@ namespace sc {
         };
         std::vector<int> get_spectra_mode(std::vector<int> indices = {}) override;
         std::vector<int> get_spectra_polarity(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_lowmz(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_highmz(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_bpmz(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_bpint(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_tic(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_rt(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_mobility(std::vector<int> indices = {}) override {
-          std::vector<double> drift;
+        std::vector<float> get_spectra_lowmz(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_highmz(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_bpmz(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_bpint(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_tic(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_rt(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_mobility(std::vector<int> indices = {}) override {
+          std::vector<float> drift;
           return drift;
         };
         std::vector<int> get_spectra_precursor_scan(std::vector<int> indices = {}) override {
           std::vector<int> precursor_scan;
           return precursor_scan;
         };
-        std::vector<double> get_spectra_precursor_mz(std::vector<int> indices = {}) override;
-        std::vector<double> get_spectra_precursor_window_mz(std::vector<int> indices = {}) override {
-          std::vector<double> precursor_window_mz;
+        std::vector<float> get_spectra_precursor_mz(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_precursor_window_mz(std::vector<int> indices = {}) override {
+          std::vector<float> precursor_window_mz;
           return precursor_window_mz;
         };
-        std::vector<double> get_spectra_precursor_window_mzlow(std::vector<int> indices = {}) override {
-          std::vector<double> precursor_window_mzlow;
+        std::vector<float> get_spectra_precursor_window_mzlow(std::vector<int> indices = {}) override {
+          std::vector<float> precursor_window_mzlow;
           return precursor_window_mzlow;
         };
-        std::vector<double> get_spectra_precursor_window_mzhigh(std::vector<int> indices = {}) override {
-          std::vector<double> precursor_window_mzhigh;
+        std::vector<float> get_spectra_precursor_window_mzhigh(std::vector<int> indices = {}) override {
+          std::vector<float> precursor_window_mzhigh;
           return precursor_window_mzhigh;
         };
-        std::vector<double> get_spectra_collision_energy(std::vector<int> indices = {}) override;
+        std::vector<float> get_spectra_collision_energy(std::vector<int> indices = {}) override;
         std::vector<int> get_polarity() override;
         std::vector<int> get_mode() override;
         std::vector<int> get_level() override;
@@ -127,10 +127,10 @@ namespace sc {
           std::vector<int> configuration;
           return configuration;
         };
-        double get_min_mz() override;
-        double get_max_mz() override;
-        double get_start_rt() override;
-        double get_end_rt() override;
+        float get_min_mz() override;
+        float get_max_mz() override;
+        float get_start_rt() override;
+        float get_end_rt() override;
         bool has_ion_mobility() override { return false; };
         MS_SUMMARY get_summary() override;
         MS_SPECTRA_HEADERS get_spectra_headers(std::vector<int> indices = {}) override;
@@ -138,9 +138,9 @@ namespace sc {
           MS_CHROMATOGRAMS_HEADERS chromatograms_headers;
           return chromatograms_headers;
         };
-        std::vector<std::vector<std::vector<double>>> get_spectra(std::vector<int> indices = {}) override;
-        std::vector<std::vector<std::vector<double>>> get_chromatograms(std::vector<int> indices = {}) override {
-          std::vector<std::vector<std::vector<double>>> chromatograms;
+        std::vector<std::vector<std::vector<float>>> get_spectra(std::vector<int> indices = {}) override;
+        std::vector<std::vector<std::vector<float>>> get_chromatograms(std::vector<int> indices = {}) override {
+          std::vector<std::vector<std::vector<float>>> chromatograms;
           return chromatograms;
         };
         MS_SPECTRUM get_spectrum(const int& idx) override;
